move types to domain
