+++ feature/links/components/cards/HorizontalCard.tsx
@@ -2,8 +2,8 @@ import { Image, View } from "react-native";
 
 import { PressableCard } from "@/components/pressable/PressableCard";
 import { ThemedText } from "@/components/text/ThemedText";
+import { type OGData } from "../../domain/models/types";
 import { useOpenBrowser } from "../../hooks/useOpenBrowser";
-import { type OGData } from "../../types/links";
 import { ErrorCard } from "./ErrorCard";
 
 type HorizontalCardProps = {
