+++ feature/links/domain/models/types/webBrowser.ts
