+++ feature/links/service/ogService.ts
@@ -1,7 +1,7 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { getLinkPreview } from "link-preview-js";
 
-import { type OGData } from "../types/links";
+import { type OGData } from "../domain/models/types";
 
 type LinkPreviewResult = {
   title?: string;
